refactor: move system handlers out of api package
